Rename `market_token_account.rs` to `create_token_accounts.rs`
